+++ src/eval.h
@@ -11,7 +11,7 @@ void _qu_eval_int(qu_parse_context *info,
 void _qu_eval_float(qu_parse_context *info,
     char *value, int interp, double *result);
 void _qu_eval_str(qu_parse_context *info,
-    char *value, int interp, char **result, int *rlen);
+    char *value, int interp, char **result, size_t *rlen);
 
 // PUBLIC API
 // Keep in sync with quire.h
@@ -23,6 +23,6 @@ void qu_node_to_int(qu_parse_context *ctx, qu_ast_node *node, uint64_t flags,
 void qu_node_to_float(qu_parse_context *ctx, qu_ast_node *node, uint64_t flags,
     double *result);
 void qu_node_to_str(qu_parse_context *ctx, qu_ast_node *node, uint64_t flags,
-    char **result, int *rlen);
+    char **result, size_t *rlen);
 
 #endif // _H_EVAL
